fixing issues after UAT

+++ src/main/java/com/dbe/domain/applicant/AppliedPersonelView_.java
@@ -7,7 +7,7 @@ import javax.persistence.metamodel.StaticMetamodel;
 public class AppliedPersonelView_ {
     public static volatile SingularAttribute<AppliedPersonelView,Long> age;
     public static volatile SingularAttribute<AppliedPersonelView,String> gender;
-    public static volatile SingularAttribute<AppliedPersonelView,Long> totalExperience;
+    public static volatile SingularAttribute<AppliedPersonelView,Double> totalExperience;
     public static volatile SingularAttribute<AppliedPersonelView,Long> vacancyId;
     public static volatile SingularAttribute<AppliedPersonelView,Double> cgpa;
     public static volatile SingularAttribute<AppliedPersonelView,Long> qualification;
